+++ src/main/java/kitchenpos/domain/Menu.java
@@ -10,6 +10,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
+import kitchenpos.product.domain.Product;
 
 @Entity
 public class Menu {
