+++ src/main/java/kitchenpos/domain/validator/TableGroupValidator.java
@@ -3,9 +3,9 @@ package kitchenpos.domain.validator;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.repository.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import org.springframework.stereotype.Component;
 
 @Component
