+++ src/main/java/kitchenpos/order/service/OrderResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.application.dto.response;
+package kitchenpos.order.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderResponse {
     private Long id;
