+++ src/main/java/kitchenpos/product/domain/ProductRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.application.product.domain;
+package kitchenpos.product.domain;
 
-import kitchenpos.application.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
