+++ src/main/java/kitchenpos/product/service/ProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.product.service;
+package kitchenpos.product.service;
 
 import java.math.BigDecimal;
