+++ src/main/java/kitchenpos/product/service/ProductResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.application.product.service;
+package kitchenpos.product.service;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.application.product.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductResponse {
     private Long id;
