+++ src/main/resources/db/migration/V1__Initialize_project_tables.sql
@@ -1,58 +1,66 @@
-CREATE TABLE orders (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    order_table_id BIGINT(20) NOT NULL,
-    order_status VARCHAR(255) NOT NULL,
-    ordered_time DATETIME NOT NULL,
+CREATE TABLE orders
+(
+    id             BIGINT(20)   NOT NULL AUTO_INCREMENT,
+    order_table_id BIGINT(20)   NOT NULL,
+    order_status   VARCHAR(255) NOT NULL,
+    ordered_time   DATETIME     NOT NULL,
     PRIMARY KEY (id)
 );
 
-CREATE TABLE order_line_item (
-    seq BIGINT(20) NOT NULL AUTO_INCREMENT,
+CREATE TABLE order_line_item
+(
+    seq      BIGINT(20) NOT NULL AUTO_INCREMENT,
     order_id BIGINT(20) NOT NULL,
-    menu_id BIGINT(20) NOT NULL,
+    menu_id  BIGINT(20) NOT NULL,
     quantity BIGINT(20) NOT NULL,
     PRIMARY KEY (seq)
 );
 
-CREATE TABLE menu (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    name VARCHAR(255) NOT NULL,
-    price DECIMAL(19, 2) NOT NULL,
-    menu_group_id BIGINT(20) NOT NULL,
+CREATE TABLE menu
+(
+    id            BIGINT(20)     NOT NULL AUTO_INCREMENT,
+    name          VARCHAR(255)   NOT NULL,
+    price         DECIMAL(19, 2) NOT NULL,
+    menu_group_id BIGINT(20)     NOT NULL,
     PRIMARY KEY (id)
 );
 
-CREATE TABLE menu_group (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
+CREATE TABLE menu_group
+(
+    id   BIGINT(20)   NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     PRIMARY KEY (id)
 );
 
-CREATE TABLE menu_product (
-    seq BIGINT(20) NOT NULL AUTO_INCREMENT,
-    menu_id BIGINT(20) NOT NULL,
+CREATE TABLE menu_product
+(
+    seq        BIGINT(20) NOT NULL AUTO_INCREMENT,
+    menu_id    BIGINT(20) NOT NULL,
     product_id BIGINT(20) NOT NULL,
-    quantity BIGINT(20) NOT NULL,
+    quantity   BIGINT(20) NOT NULL,
     PRIMARY KEY (seq)
 );
 
-CREATE TABLE order_table (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    table_group_id BIGINT(20),
-    number_of_guests INT(11) NOT NULL,
-    empty BIT(1) NOT NULL,
+CREATE TABLE order_table
+(
+    id               BIGINT(20) NOT NULL AUTO_INCREMENT,
+    table_group_id   BIGINT(20),
+    number_of_guests INT(11)    NOT NULL,
+    empty            BIT(1)     NOT NULL,
     PRIMARY KEY (id)
 );
 
-CREATE TABLE table_group (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    created_date DATETIME NOT NULL,
+CREATE TABLE table_group
+(
+    id           BIGINT(20) NOT NULL AUTO_INCREMENT,
+    created_date DATETIME   NOT NULL,
     PRIMARY KEY (id)
 );
 
-CREATE TABLE product (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    name VARCHAR(255) NOT NULL,
+CREATE TABLE product
+(
+    id    BIGINT(20)     NOT NULL AUTO_INCREMENT,
+    name  VARCHAR(255)   NOT NULL,
     price DECIMAL(19, 2) NOT NULL,
     PRIMARY KEY (id)
 );
