+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -11,26 +11,27 @@ import java.util.List;
 import kitchenpos.SpringBootTestWithProfiles;
 import kitchenpos.application.dto.request.MenuProductRequest;
 import kitchenpos.application.dto.request.MenuRequest;
-import kitchenpos.application.dto.request.OrderRequest;
-import kitchenpos.application.dto.request.OrderRequest.OrderLineItemRequest;
-import kitchenpos.application.dto.request.OrderStatusRequest;
-import kitchenpos.application.product.service.ProductRequest;
+import kitchenpos.order.service.OrderRequest;
+import kitchenpos.order.service.OrderRequest.OrderLineItemRequest;
+import kitchenpos.order.service.OrderStatusRequest;
+import kitchenpos.order.service.OrderService;
+import kitchenpos.product.service.ProductRequest;
 import kitchenpos.application.dto.request.TableRequest;
 import kitchenpos.application.dto.response.MenuGroupResponse;
 import kitchenpos.application.dto.response.MenuResponse;
-import kitchenpos.application.dto.response.OrderResponse;
+import kitchenpos.order.service.OrderResponse;
 import kitchenpos.application.dto.response.OrderTableResponse;
-import kitchenpos.application.product.service.ProductResponse;
-import kitchenpos.application.product.service.ProductService;
+import kitchenpos.product.service.ProductResponse;
+import kitchenpos.product.service.ProductService;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.domain.repository.MenuGroupRepository;
 import kitchenpos.domain.repository.MenuProductRepository;
 import kitchenpos.domain.repository.MenuRepository;
 import kitchenpos.domain.repository.OrderLineItemRepository;
-import kitchenpos.domain.repository.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.domain.repository.OrderTableRepository;
-import kitchenpos.application.product.domain.ProductRepository;
+import kitchenpos.product.domain.ProductRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
