+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -7,10 +7,10 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.SpringBootTestWithProfiles;
-import kitchenpos.application.product.service.ProductRequest;
-import kitchenpos.application.product.service.ProductResponse;
-import kitchenpos.application.product.service.ProductService;
-import kitchenpos.application.product.domain.ProductRepository;
+import kitchenpos.product.service.ProductRequest;
+import kitchenpos.product.service.ProductResponse;
+import kitchenpos.product.service.ProductService;
+import kitchenpos.product.domain.ProductRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
