+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -13,10 +13,10 @@ import kitchenpos.SpringBootTestWithProfiles;
 import kitchenpos.application.dto.request.TableGroupRequest;
 import kitchenpos.application.dto.request.TableGroupRequest.OrderTableId;
 import kitchenpos.application.dto.response.TableGroupResponse;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.repository.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.domain.repository.OrderTableRepository;
 import kitchenpos.domain.repository.TableGroupRepository;
 import org.junit.jupiter.api.AfterEach;
